requirements.txt
# Backend
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
alembic
passlib[bcrypt]
python-jose
python-multipart

# Scraping y validación
beautifulsoup4
requests
phonenumbers
scraperapi-sdk
email-validator>=2.1.0

# IA
openai

# Stripe
stripe

# Frontend
streamlit>=1.31
streamlit_js_eval
extra-streamlit-components

# Data
pandas

# Testing
pytest
psycopg2-binary
testcontainers
pytest-asyncio

